Use EpidAttestationVerificationBuilder
